+++ test/registration/support/registered_server.ex
@@ -4,7 +4,8 @@ defmodule Commanded.Registration.RegisteredServer do
 
   def start_link(args) do
     name = args[:name]
-    GenServer.start_link(__MODULE__, name, [name: name])
+
+    GenServer.start_link(__MODULE__, name, name: name)
   end
 
   def init(name), do: {:ok, name}