+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add `.formatter.exs` to Hex package ([#247](https://github.com/commanded/commanded/pull/247)).
 - Event upcasting ([#263](https://github.com/commanded/commanded/pull/263)).
 - Support `:ok` tagged tuple events from aggregate ([#268](https://github.com/commanded/commanded/pull/268)).
+- Modify `Commanded.Registration.start_child/3` to pass a child_spec ([#273](https://github.com/commanded/commanded/pull/273)).
 
 ## v0.18.0